+++ packages/orq-rc/RELEASES.md
@@ -8239,3 +8239,13 @@ Based on:
 - [python v4.10.0-rc.49] packages/orq-rc
 ### Releases
 - [PyPI v4.10.0-rc.49] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.49 - packages/orq-rc
+
+## 2026-05-28 08:19:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.10.0-rc.50] packages/orq-rc
+### Releases
+- [PyPI v4.10.0-rc.50] https://pypi.org/project/orq-ai-sdk/4.10.0-rc.50 - packages/orq-rc